Remove separate custom level array, other cleanup
Instead of putting custom levels inside of a separate array and
element, it appends them to the Simple or Compound level array
depending on the quaver (for now).

Also removes level 8 which was supposed to be a combination of
the three compound levels, but this already was manually created
in level 7.

Other miscellaneous cleanup.