Handle valid and invalid param blocks gracefully
Allows blocks entered in via query params to be handled gracefully
even if they are not valid or they exist in an incorrect level.

Note that Levels are given precedence over Blocks. A block that
exists in another level will simply be ignored if it is not
available in the current level.