textNote Counts working for simple meter
textNote-based count voices are implemented with the correct
behavior in simple meter only. This is done by iterating over
each `smallest_duration` (16th note by default) in a separate
voice and putting a text note with text on each beat and a
blank textNote on non-beat 16ths.

This does not currently work for triplets nor does it work
for complex meter yet.