Improved triplet counts formatting
Passes any triplet tickables directly to the countsVoice to ensure
that tick contexts line up properly. Drawing the triplets is still
somewhat strange, but that seems to be related to the workings of
the library itself.

Also, need to handle the annotation on the first of the triplets
if it resides on the beat.