Implements Difficulty Schema
When a non-custom difficulty is selected, the blocks with the
highest value of base difficulty are chosen more often.

Base difficulty is calculated by the character code of the lowest
difficulty when sorted by character code. So, for example, a
block that is included in difficulties of A, C, and F will have a
base difficulty of A (or 1).

This is not applied to custom rhythm sets (when sets are modified by
selecting individual blocks.