Use Docs Visualization API to retrieve Spreadsheet data
The Google Sheets API recently changed to disallow CORS requests to
retrieve spreadsheet data, even for those sheets that are publically
shared and streaming. This change uses the Google Visualization API
to retrieve a function that includes the data and just rips the data
out from the function for use in the app.

This is an admitedly brittle solution that gets the functionality back
up for now.